+++ grid2op/Action/__init__.py
@@ -1,31 +1,56 @@
 __all__ = [
+    # Internals
     "BaseAction",
     "ActionSpace",
-    "PowerLineSet",
     "SerializableActionSpace",
-    "TopoAndRedispAction",
-    "TopologyAction",
+    "PlayableAction",
     "VoltageOnlyAction",
-    'DontAct',
-    "HelperAction",
     "CompleteAction",
+    # Usable
+    "DontAct",
+    "PowerlineSetAction",
+    "PowerlineChangeAction",
+    "TopologyAction",
+    "TopologyAndDispatchAction",
+    "TopologySetAction",
+    "TopologySetAndDispatchAction",
+    "TopologyChangeAction",
+    "TopologyChangeAndDispatchAction",
+    # Backwards compat
+    "TopoAndRedispAction",
+    "HelperAction",
     "Action"
 ]
 
+# Internals
 from grid2op.Action.BaseAction import BaseAction
 from grid2op.Action.ActionSpace import ActionSpace
-from grid2op.Action.PowerLineSet import PowerLineSet
 from grid2op.Action.SerializableActionSpace import SerializableActionSpace
-from grid2op.Action.TopoAndRedispAction import TopoAndRedispAction
-from grid2op.Action.TopologyAction import TopologyAction
+from grid2op.Action.PlayableAction import PlayableAction
 from grid2op.Action.VoltageOnlyAction import VoltageOnlyAction
-from grid2op.Action.DontAct import DontAct
-import warnings
 
+from grid2op.Action.DontAct import DontAct
+from grid2op.Action.PowerlineSetAction import PowerlineSetAction
+from grid2op.Action.PowerlineChangeAction import PowerlineChangeAction
+from grid2op.Action.TopologyAction import TopologyAction
+from grid2op.Action.TopologyAndDispatchAction import TopologyAndDispatchAction
+from grid2op.Action.TopologySetAction import TopologySetAction
+from grid2op.Action.TopologySetAndDispatchAction import TopologySetAndDispatchAction
+from grid2op.Action.TopologyChangeAction import TopologyChangeAction
+from grid2op.Action.TopologyChangeAndDispatchAction import TopologyChangeAndDispatchAction
+
+import warnings
 
 # CompleteAction to be symetrical to CompleteObservation
 CompleteAction = BaseAction
 
+class TopoAndRedispAction(TopologyAndDispatchAction):
+    def __init__(self, gridobj):
+        super().__init__(gridobj)
+        warnings.warn("TopoAndRedispAction has been renamed to TopologyAndDispatchAction"
+                      " -- The old name will be removed in future versions",
+                      category=PendingDeprecationWarning)
+    
 
 class HelperAction(ActionSpace):
     def __init__(self, *args, **kwargs):