+++ grid2op/data/rte_case14_realistic/config.py
@@ -1,13 +1,13 @@
-from grid2op.Action import TopoAndRedispAction
+from grid2op.Action import TopologyAndDispatchAction
 from grid2op.Reward import RedispReward
 from grid2op.Rules import DefaultRules
 from grid2op.Chronics import Multifolder
-from grid2op.Chronics import GridStateFromFileWithForecasts
+from grid2op.Chronics import GridStateFromFile
 from grid2op.Backend import PandaPowerBackend
 
 config = {
     "backend": PandaPowerBackend,
-    "action_class": TopoAndRedispAction,
+    "action_class": TopologyAndDispatchAction,
     "observation_class": None,
     "reward_class": RedispReward,
     "gamerules_class": DefaultRules,